readme.md 작성

+++ src/components/Todo.tsx
@@ -10,7 +10,8 @@ const Todo = ({ todo }: Props) => {
   const { updateTodo, deleteTodo } = useTodos();
 
   const deleteTodoObj = (id: string) => {
-    deleteTodo(id);
+    const userConfirmed = window.confirm("정말 삭제하시겠습니까?");
+    if (userConfirmed) deleteTodo(id);
   };
 
   const switchTodoObj = (id: string) => {
